+++ ddtrace/appsec/iast/taint_sinks/command_injection.py
@@ -1,10 +1,24 @@
+import re
 from typing import List
+from typing import TYPE_CHECKING
 from typing import Union
 
+import six
+
 from ddtrace.appsec.iast import oce
+from ddtrace.appsec.iast._utils import _scrub_get_tokens_positions
 from ddtrace.appsec.iast.constants import EVIDENCE_CMDI
 from ddtrace.appsec.iast.constants import VULN_CMDI
 from ddtrace.appsec.iast.taint_sinks._base import VulnerabilityBase
+
+
+if TYPE_CHECKING:
+    from typing import Any
+    from typing import Dict
+
+    from ddtrace.appsec.iast.reporter import Vulnerability
+
+_INSIDE_QUOTES_REGEXP = re.compile(r"(?:\s|^)(\S*\/\S*)")
 
 
 @oce.register
@@ -19,6 +33,20 @@ class CommandInjection(VulnerabilityBase):
 
             evidence_value, sources = taint_ranges_as_evidence_info(evidence_value)
         super(CommandInjection, cls).report(evidence_value=evidence_value, sources=sources)
+
+    @classmethod
+    def _extract_sensitive_tokens(cls, vulns_to_text):
+        # type: (Dict[Vulnerability, str]) -> Dict[int, Dict[str, Any]]
+
+        ret = {}  # type: Dict[int, Dict[str, Any]]
+        for vuln, text in six.iteritems(vulns_to_text):
+            vuln_hash = hash(vuln)
+            ret[vuln_hash] = {
+                "tokens": set(_INSIDE_QUOTES_REGEXP.findall(text)),
+            }
+            ret[vuln_hash]["token_positions"] = _scrub_get_tokens_positions(text, ret[vuln_hash]["tokens"])
+
+        return ret
 
 
 def _iast_report_cmdi(shell_args):
