+++ ddtrace/appsec/iast/taint_sinks/sql_injection.py
@@ -25,7 +25,7 @@ _INSIDE_QUOTES_REGEXP = re.compile(r'["\']([^"\']*?)["\']')
 class SqlInjection(VulnerabilityBase):
     vulnerability_type = VULN_SQL_INJECTION
     evidence_type = EVIDENCE_SQL_INJECTION
-    scrub_evidence = True
+    _redacted_pattern = True
 
     @classmethod
     def report(cls, evidence_value=None, sources=None):
